// src/role/enum/permissions.enum.ts
export enum PermissionType {
  //ROLE AND PERMISSIONS
  MANAGE_ROLE = 'MANAGE_ROLE',
  CREATE_ROLE = 'CREATE_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',
  UPDATE_ROLE = 'UPDATE_ROLE',
  GET_ALL_ROLE = 'GET_ALL_ROLE',
  ASSIGN_ROLE = 'ASSIGN_ROLE',
  UNASSIGN_ROLE = 'UNASSIGN_ROLE',

  //PRODUCT
  MANAGE_PRODUCT = 'MANAGE_PRODUCT',
  CREATE_PRODUCT = 'CREATE_PRODUCT',
  DELETE_PRODUCT = 'DELETE_PRODUCT',
  UPDATE_PRODUCT = 'UPDATE_PRODUCT',
  GET_ALL_PRODUCT = 'GET_ALL_PRODUCT',

  //EVENT
  MANAGE_EVENT = 'MANAGE_EVENT',
  CREATE_EVENT = 'CREATE_EVENT',
  DELETE_EVENT = 'DELETE_EVENT',
  UPDATE_EVENT = 'UPDATE_EVENT',
  GET_ALL_EVENT = 'GET_ALL_EVENT',
  GET_EVENT = 'GET_EVENT',
}